% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGT_par.R
\name{SGT_par}
\alias{SGT_par}
\title{Compute snow gamma parameters per elevation zone}
\usage{
SGT_par(path_dtm, shp, station, path_BestPar, ar50f, ar50w)
}
\arguments{
\item{shp}{single catchment polygon feature (as shapefile or sf)}

\item{station}{station number, finishin with ".0". F eg "7.23.0"}

\item{ar50f}{forest & wetlands from AR50 dataset (artype=30 & 60)}

\item{ar50w}{water bodies from AR50 dataset (artype=81)}

\item{path}{path with the raster catalog}

\item{path_Bestpar}{path with catchment parameter file to extract the elevation zone breaks}
}
\value{
A table with all elevation zones, Std_SqS, percentage of vegetation, and gamma parameters a0 & D
}
\description{
Clip raster to polygon. It uses a shapefile (or sf) as input, locate the extent on the raster catalog, crops the raster, then clip
}
\author{
Luis Barreiro
}
