% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_DDD.R
\name{stat_DDD}
\alias{stat_DDD}
\title{Compute snow gamma parameters per elevation zone}
\usage{
stat_DDD(station, dtm_res, path_DDD_R, path_DDD_T, path_MODIS, cloud = 30)
}
\arguments{
\item{station}{station number, finishin with ".0". F eg "7.23.0"}

\item{dtm_res}{dtm resolution in meter for accessing the right folder of results (1/10)}

\item{path_DDD_R}{path to DDD output using SG_R}

\item{path_DDD_T}{path to DDD output using SG_T}

\item{path_MODIS}{path to Modis files with SCA observations}

\item{cloud}{maximum cloud cover alq loud for using Modis SCA (default=30)}
}
\value{
A table with all results, efficiency creiteria, biases and whether they are significant or not. Q_KGE_cal and Q_KGE_val are the efficiency criteria Kling-Gupta Efficiency (KGE) (Gupta et al. (2009) and Kling et al. (2012)) for calibration and validation dataset respectively, and for SG_R (_R) and SG_T (_T). Q_NSE_cal and Q_NSE_val are the Nash-Sutcliffe efficiency (NSE) criterion (Nash & Sutcliffe, 1970) for calibration and validation. Q_bias_cal and Q_bias_val are the bias between observed and estimated streamflow. Qp_bias_R and Qp_bias_T are the (maximum annual peak flow for spring season) bias obs/estimated for SG_R and SG_T. SWE_bias_cal and SWE_bias_val are the biases between SG_R and SG_T. SCA_avg is the average SCA for the melting season for mMODIS (_modis), SG_R (_R) and SG_T (_T). SCA_bias_R and SCA_bias_T are the SCA bias obs/estimated for SG_R and SC_T. For all bias calculations, the symbol (*) marks when it is significant (sig. level 5%)
}
\description{
Clip raster to polygon. It uses a shapefile (or sf) as input, locate the extent on the raster catalog, crops the raster, then clip
}
\author{
Luis Barreiro
}
